pyproject.toml
[tool.poetry]
name = "langgraph-fin-agent"
version = "0.1.0"
description = "Finance agent built using Langgraph"
authors = ["Ankur Duggal <aduggal@arize.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
langgraph = "^0.2.45"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.8"
beautifulsoup4 = "^4.13.3"
openinference-instrumentation-langchain = "^0.1.35"
arize-otel = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
finchat = "langgraph_fin_agent.main:main"